Removed import sys. Added new command option --System
Packages sys was used in earlier versions to provide argv values.
With the switch to argparse, rolling our own command line parsing
is unneccessary. Removed as it was not used.

--System has been added to provide the ability to choose from a
wholy random run, or one that starts with the same seed. The
default seed is 42, and will be used if --System is not specified.