Limited the number of chronons a simulation can run.
Wa-Tor would run until either all fishes or sharks were dead. This is
nice from a aesthic or philosophical point of view; but with the
storing of images for animation later, impractical for at least two
reasons: space and time. If an set length animation is wanted, there is
no point in having to watch the program run, then terminate it once enough
frames have been generated. Secondly, files take up space and it is
possible to consume to much of the disc without limiting the number
generated.

This change implements both a default (999999) and optional limit. The
default is more than 11 hours at 24fps, so it should suffice for long
runs.

NOTE: creating and storing files should be addressed in a later release.