Prepare class Sea for save and restore
Earlier attempts to save a restore tried to save the entire object,
instead we will now just save the state then recreate the sea at
the time of restoration.

exportSea() returns a sea's parameters as a tuple.
setFileNumber() will allow setting filenumber to the last used.
setCreatureTag0 will allow setting creatureTag to the last used.

To restore the following will be done:
	Create a sea the saved sea's dimensions.
	Move the creatureTag, and filenumber ahead to the last used.
	Add all the creatures to the sea.