Implemented save-restore
In place of exporting the class Sea, export the parameters of the sea,
and its creatures. Restore by reinitializing from these parameters.

The current chronon, and the random number generator's state
are not saved. Saving them would allow the log to pick up from the
last saved chronon instead of starting for zero again. Saving the
RNG might make things smoother when transitioning to a restart, but
the image does not change dramatically if the same RNG type is used.
Saving the type of RNG, is a potential future revision.