Restore now resets the age and starve values
Saving a creature did save age and starve, but these were not
reset when a world was recovered. This had all creatures starting
from age zero. This might have lead to a skew in the simulation, as
creatures are now younger. For example, sharks that might die of
starvation were given a longer time to live.