Vary initial ages of creatures
This addresses issue #7 "Vary age of sharks when sea is initialized"
With all creatures, shark and fish, being born at initialization,
the spawning of both sharks and fishes, and sharks dying of starvation
occurs in lockstep. This gives some simulations a sawtooth looking
shark populations. Varying the creature ages is intended to smoooth
this out.