Vary ages of creatures at initialization
Although this was implemented, it was done in such a way that
all creatures were given random ages, and starvation values when
they were born. This was good for initialization, but not fair
when spawning. Revised the code to make random values occur during
the generation of the sea, and at no other time.

Shark.__str__(), and Fish.__str__() were changed to allow insight
into creature behavior.