Spawning now is no p=0.30 once spawing age has been reached
The sawtooth pattern in shark populations is eased by spreading out
the age at which a creature spawns, this prevents the all or nothing
aspect of a cohort when reproducing. If the simulation runs for many
chronons, a jagged pattern does emerge, but it is much reduced from
the original sawtooth.